# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DuitPlan API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  # Onboarding
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions with filters
      parameters:
        - name: month
          in: query
          schema:
            type: ["string", "null"]
          description: "Filter by month (YYYY-MM format)"
        - name: categoryId
          in: query
          schema:
            type: ["string", "null"]
        - name: accountId
          in: query
          schema:
            type: ["string", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
          description: "debit or credit"
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a manual transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionBody"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Budgets
...
  /uploads:
    post:
      operationId: uploadStatement
      tags: [uploads]
      summary: Upload a bank statement PDF
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadStatementBody"
      responses:
        "201":
          description: Document uploaded and parsed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedDocument"
  /uploads/{id}/rows:
    get:
      operationId: getImportedRows
      tags: [uploads]
      summary: Get parsed transaction rows for review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Parsed rows for review
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImportedTransactionRow"
  /uploads/{id}/confirm:
    post:
      operationId: confirmImport
      tags: [uploads]
      summary: Confirm import of reviewed transaction rows
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmImportBody"
      responses:
        "200":
          description: Import confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportConfirmResult"

  # Subscription
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Auth schemas
...
    Profile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        fullName:
          type: string
        currency:
          type: string
        payday:
          type: integer
        monthlyIncome:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - fullName
        - currency
        - payday
        - monthlyIncome
        - createdAt
        - updatedAt
    UpdateProfileBody:
      type: object
      properties:
        fullName:
          type: ["string", "null"]
        payday:
          type: ["integer", "null"]
        monthlyIncome:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]

    # Onboarding
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        accountId:
          type: ["string", "null"]
        categoryId:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
        date:
          type: string
        amount:
          type: string
        type:
          type: string
        description:
          type: string
        merchant:
          type: ["string", "null"]
        source:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - date
        - amount
        - type
        - description
        - source
        - createdAt
        - updatedAt
    CreateTransactionBody:
      type: object
      properties:
        date:
          type: string
        amount:
          type: string
        type:
          type: string
        description:
          type: string
        merchant:
          type: ["string", "null"]
        categoryId:
          type: ["string", "null"]
        accountId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - date
        - amount
        - type
        - description
    UpdateTransactionBody:
      type: object
      properties:
        date:
          type: ["string", "null"]
        amount:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        merchant:
          type: ["string", "null"]
        categoryId:
          type: ["string", "null"]
        accountId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # Budget
...
    UploadStatementBody:
      type: object
      properties:
        bankType:
          type: string
        file:
          type: string
          format: binary
      required:
        - bankType
        - file
    UploadedDocument:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        fileName:
          type: string
        storagePath:
          type: string
        bankType:
          type: string
        parseStatus:
          type: string
        uploadedAt:
          type: string
        rowCount:
          type: integer
      required:
        - id
        - userId
        - fileName
        - storagePath
        - bankType
        - parseStatus
        - uploadedAt
        - rowCount
    ImportedTransactionRow:
      type: object
      properties:
        id:
          type: string
        uploadedDocumentId:
          type: string
        rawDate:
          type: ["string", "null"]
        rawDescription:
          type: ["string", "null"]
        rawAmount:
          type: ["string", "null"]
        normalizedDate:
          type: ["string", "null"]
        normalizedDescription:
          type: ["string", "null"]
        normalizedAmount:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        categorySuggestion:
          type: ["string", "null"]
        confidence:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - uploadedDocumentId
        - status
        - createdAt
    ConfirmImportBody:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ConfirmImportRow"
      required:
        - rows
    ConfirmImportRow:
      type: object
      properties:
        rowId:
          type: string
        categoryId:
          type: ["string", "null"]
        skip:
          type: boolean
      required:
        - rowId
        - skip
    ImportConfirmResult:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
      required:
        - imported
        - skipped

    # Subscription
...