# lib/api-zod/src/generated/types/importConfirmResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DuitPlan API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ImportConfirmResult {
  imported: number;
  skipped: number;
}